server/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
python-dotenv
pillow
werkzeug
requests
python-multipart
openai
